[Ui-CustomView] ConfirmButton

+++ app/src/main/java/com/ray/template/common/util/ViewUtil.kt
@@ -3,11 +3,11 @@ package com.ray.template.common.util
 import android.content.res.Resources
 import android.util.TypedValue
 
-val Int.dp: Int
+val Int.dp: Float
     get() {
         return TypedValue.applyDimension(
             TypedValue.COMPLEX_UNIT_DIP,
             this.toFloat(),
             Resources.getSystem().displayMetrics
-        ).toInt()
+        )
     }
